Move bigsky util package to util/cliutil since lots of stuff uses it
