+++ cmd/fakermaker/main.go
@@ -10,8 +10,8 @@ import (
 	"os"
 	"runtime"
 
-	cliutil "github.com/bluesky-social/indigo/cmd/gosky/util"
 	"github.com/bluesky-social/indigo/fakedata"
+	"github.com/bluesky-social/indigo/util/cliutil"
 	"github.com/bluesky-social/indigo/version"
 
 	_ "github.com/joho/godotenv/autoload"
