+++ pds/handlers_test.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/bluesky-social/indigo/api/atproto"
 	"github.com/bluesky-social/indigo/carstore"
-	cliutil "github.com/bluesky-social/indigo/cmd/gosky/util"
 	"github.com/bluesky-social/indigo/plc"
+	"github.com/bluesky-social/indigo/util/cliutil"
 	"github.com/whyrusleeping/go-did"
 	"gorm.io/gorm"
 )
