+++ util/cliutil/key_test.go
